+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-25T04:58:43.483Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-25T12:44:28.109Z")
     r"""The date and time the resource was last updated"""