+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAWP23KWBHXB4MBED0HAHKBS"
+        "01KAXGPT5A8N6ZEGGGG9PA4FWG"
     )
     r"""The unique identifier of the data source"""