+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -250,7 +250,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-25T04:58:43.483Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-25T12:44:28.109Z")
     r"""The date and time the resource was last updated"""