+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAWP23KV1FJ6C9F6VHT2MFET"
+        "01KAXGPT59GBDFGT1WCT9NCAWY"
     )
     r"""The unique identifier of the data source"""