+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAWP23KZ170RKZQQXF821AB4"
+        "01KAXGPT5DSH216AD9N652A9VC"
     )
     r"""The unique identifier of the data source"""