+++ packages/orq-rc/src/orq_ai_sdk/responses.py
@@ -85,7 +85,7 @@ class Responses(BaseSDK):
             base_url=base_url,
             url_variables=url_variables,
             request=request,
-            request_body_required=False,
+            request_body_required=True,
             request_has_path_params=True,
             request_has_query_params=True,
             user_agent_header="user-agent",
@@ -95,9 +95,9 @@ class Responses(BaseSDK):
             get_serialized_body=lambda: utils.serialize_request_body(
                 request.request_body,
                 False,
-                True,
+                False,
                 "json",
-                Optional[models.CreateAgentResponseRequestRequestBody],
+                models.CreateAgentResponseRequestRequestBody,
             ),
             timeout_ms=timeout_ms,
         )
@@ -213,7 +213,7 @@ class Responses(BaseSDK):
             base_url=base_url,
             url_variables=url_variables,
             request=request,
-            request_body_required=False,
+            request_body_required=True,
             request_has_path_params=True,
             request_has_query_params=True,
             user_agent_header="user-agent",
@@ -223,9 +223,9 @@ class Responses(BaseSDK):
             get_serialized_body=lambda: utils.serialize_request_body(
                 request.request_body,
                 False,
-                True,
+                False,
                 "json",
-                Optional[models.CreateAgentResponseRequestRequestBody],
+                models.CreateAgentResponseRequestRequestBody,
             ),
             timeout_ms=timeout_ms,
         )