+++ packages/orq-rc/RELEASES.md
@@ -4209,3 +4209,13 @@ Based on:
 - [python v4.1.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v4.1.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.3 - packages/orq-rc
+
+## 2025-11-25 12:46:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.656.1 (2.753.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.1.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.5 - packages/orq-rc